feat: move the refid global out of the openai module
